# docs/smart-finance-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Finance API",
    "version": "0.0.3",
    "description": "Smart Finance uses a JSESSIONID cookie and CSRF protection.\nFirst GET /api/auth/csrf and retain its cookie. Send the returned token in\nX-CSRF-TOKEN for registration, login and logout. Registration creates an\naccount without signing in; log in separately and refresh CSRF after login.\n"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "operationId": "getHealth",
        "summary": "检查应用是否能够响应",
        "description": "返回 status=UP，确认 HTTP 服务正在响应；不检查数据库或其他外部依赖。",
        "responses": {
          "200": {
            "description": "Application is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/csrf": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "getCsrf",
        "summary": "获取当前会话的 CSRF 令牌",
        "description": "返回令牌及其请求头和参数名称；同时保留会话 Cookie，登录成功后重新获取令牌。",
        "responses": {
          "200": {
            "description": "CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "register",
        "summary": "注册账号",
        "description": "创建账号，不自动登录。username 为 1–64 个非空白字符；password 至少\n8 个字符、最多 72 个 UTF-8 字节，且不能全为空白；confirmPassword\n必须与 password 完全一致。密码仅以 BCrypt 哈希保存。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "账号创建成功，返回用户名；之后需调用登录接口",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "用户名已被注册（USERNAME_TAKEN），不会覆盖原账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "login",
        "summary": "校验数据库账号并建立登录会话",
        "description": "使用用户名查询数据库，通过 BCrypt 校验密码并检查账号启用状态。\n请求必须携带同一会话的 Cookie 和 CSRF 请求头。成功后轮换 JSESSIONID，\n返回用户名和身份，并保存服务端登录状态；浏览器需保留新 Cookie 并刷新 CSRF 令牌。\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in; retain the rotated JSESSIONID cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "getCurrentUser",
        "summary": "查询当前登录用户",
        "description": "返回当前会话的用户名和数据库最新身份，不返回密码哈希；未登录或会话失效时返回 401。",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "logout",
        "summary": "退出登录并销毁会话",
        "description": "要求有效会话和当前 CSRF 令牌；销毁服务端会话并清除认证信息，刷新页面后仍需重新登录。",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "responses": {
          "204": {
            "description": "Session invalidated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "listUsers",
        "summary": "管理员查询用户列表",
        "description": "仅 ADMIN 可访问；返回身份和启用状态，不返回密码或密码哈希。",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSummary"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/admin/users/{id}/role": {
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "updateUserRole",
        "summary": "管理员修改用户身份",
        "description": "仅 ADMIN 可操作，要求 Session Cookie 和 CSRF 请求头。\n身份写入数据库，目标用户下一次请求即使用新权限。\n不允许将最后一个启用的管理员降级为一般用户。\n",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功，返回最新用户身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "用户不存在（USER_NOT_FOUND）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "不能降级最后一个启用的管理员（LAST_ADMIN）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "X-CSRF-TOKEN",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "Missing, malformed or invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Invalid credentials or unauthenticated session",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Missing or invalid CSRF token or insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN",
          "USER"
        ],
        "description": "ADMIN 为管理员，USER 为一般用户"
      },
      "RoleUpdateRequest": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "UserSummary": {
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "enabled"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password",
          "confirmPassword"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "pattern": "\\S+",
            "description": "用户名，不允许包含空白"
          },
          "password": {
            "type": "string",
            "format": "password",
            "writeOnly": true,
            "minLength": 8,
            "maxLength": 72,
            "description": "至少 8 个字符，最多 72 个 UTF-8 字节，不能全为空白"
          },
          "confirmPassword": {
            "type": "string",
            "format": "password",
            "writeOnly": true,
            "minLength": 8,
            "maxLength": 72,
            "description": "必须与 password 完全一致"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "pattern": "\\S+"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 1,
            "maxLength": 72
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "username",
          "role"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "CsrfResponse": {
        "type": "object",
        "required": [
          "token",
          "headerName",
          "parameterName"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "headerName": {
            "type": "string"
          },
          "parameterName": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}